Update master from dev

+++ AdoptSPRING/src/main/java/fr/adoptunstage/spring/services/OffreService.java
@@ -1,7 +1,6 @@
 package fr.adoptunstage.spring.services;
 
 import java.util.HashSet;
-import java.util.Optional;
 import java.util.Set;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -16,17 +15,22 @@ import fr.adoptunstage.spring.message.request.SignUpFormOffre;
 import fr.adoptunstage.spring.message.request.SignUpPostuler;
 import fr.adoptunstage.spring.message.response.ResponseMessage;
 import fr.adoptunstage.spring.models.Entreprise;
-import fr.adoptunstage.spring.models.HTMLMail;
+import fr.adoptunstage.spring.models.EntrepriseMail;
 import fr.adoptunstage.spring.models.Offre;
 import fr.adoptunstage.spring.models.Stagiaire;
+import fr.adoptunstage.spring.models.StagiaireMail;
 import fr.adoptunstage.spring.repos.OffreRepository;
 import fr.adoptunstage.spring.repos.UserRepository;
+import fr.adoptunstage.spring.security.services.AuthenticationUser;
 
 @Service
 public class OffreService {
 
 	@Autowired
 	MailService mailRepository;
+	
+	@Autowired
+	AuthenticationUser authenticationUser;
 
 	@Autowired
 	OffreRepository repository;
@@ -77,16 +81,39 @@ public class OffreService {
 
 		Offre offre = repository.findById(id)
 				.orElseThrow(() -> new UsernameNotFoundException("Offre Not Found with -> id : " + id));
-		;
+		
 		offre.getEntreprise().setPassword("");
-		;
+		
 		return offre;
+
+	}
+	
+	public ResponseEntity<?> getModifierOffre(long id) {
+
+		Offre offre = repository.findById(id)
+				.orElseThrow(() -> new UsernameNotFoundException("Offre Not Found with -> id : " + id));
+		String usernameEntreprise = offre.getEntreprise().getUsername();
+		
+		if (authenticationUser.isValidate(usernameEntreprise)) {
+			offre.getEntreprise().setPassword("");		
+			return new ResponseEntity<>(offre, HttpStatus.OK);
+		}
+		return new ResponseEntity<>("Pas le bon utilisateur", HttpStatus.FORBIDDEN);
 
 	}
 
 	public ResponseEntity<String> deleteOffre(@PathVariable("id") long id) {
-		repository.deleteById(id);
-		return new ResponseEntity<>("offre a été supprimée !", HttpStatus.OK);
+		Offre offre = repository.findById(id)
+				.orElseThrow(() -> new UsernameNotFoundException("Offre Not Found with -> id : " + id));
+		String usernameEntreprise = offre.getEntreprise().getUsername();
+		
+		if (authenticationUser.isValidate(usernameEntreprise)) {
+			repository.deleteById(id);
+			return new ResponseEntity<>("offre a été supprimée !", HttpStatus.OK);
+		}		
+		else {
+			return new ResponseEntity<>("Pas le bon utilisateur", HttpStatus.FORBIDDEN);
+		}	
 	}
 
 	public ResponseEntity<String> deleteAll() {
@@ -94,25 +121,27 @@ public class OffreService {
 		return new ResponseEntity<>("Toutes les offres ont été supprimé!", HttpStatus.OK);
 	}
 
-	public ResponseEntity<Offre> updateOffre(@PathVariable("id") long id, @RequestBody Offre offre) {
-		Optional<Offre> offreData = repository.findById(id);
-
-		if (offreData.isPresent()) {
-			Offre _offre = offreData.get();
-			_offre.setTitre(offre.getTitre());
-			_offre.setDescription(offre.getDescription());
-			_offre.setDateDebut(offre.getDateDebut());
-			_offre.setDateFin(offre.getDateFin());
-			_offre.setRue(offre.getRue());
-			_offre.setVille(offre.getVille());
-			_offre.setCodePostal(offre.getCodePostal());
-			_offre.setActive(offre.isActive());
-
-			return new ResponseEntity<>(repository.save(_offre), HttpStatus.OK);
-		} else {
-
-			return new ResponseEntity<>(HttpStatus.NOT_FOUND);
-		}
+	public ResponseEntity<?> updateOffre(@PathVariable("id") long id, @RequestBody Offre offre) {
+		Offre _offre = repository.findById(id).orElseThrow(
+				() -> new UsernameNotFoundException("Offre avec l'id suivant introuvable : " + id));;
+		
+			String usernameEntreprise = _offre.getEntreprise().getUsername();
+			
+			if (authenticationUser.isValidate(usernameEntreprise)) {	
+				_offre.setTitre(offre.getTitre());
+				_offre.setDescription(offre.getDescription());
+				_offre.setDateDebut(offre.getDateDebut());
+				_offre.setDateFin(offre.getDateFin());
+				_offre.setRue(offre.getRue());
+				_offre.setVille(offre.getVille());
+				_offre.setCodePostal(offre.getCodePostal());
+				_offre.setActive(offre.isActive());
+	
+				return new ResponseEntity<>(repository.save(_offre), HttpStatus.OK);
+			}
+			
+			else {return new ResponseEntity<>(HttpStatus.FORBIDDEN);}
+		
 	}
 
 	public ResponseEntity<?> postOffre(String username, SignUpFormOffre requestOffre) {
@@ -141,29 +170,35 @@ public class OffreService {
 
 		offre.setStagiaire(stagiaire);
 
+		EntrepriseMail mailToEntreprise = new EntrepriseMail(offre.getEntreprise().getContactMail(), offre.getTitre(),
+				requestPostuler.getMotivation(), stagiaire.getPrenom(), stagiaire.getName(), stagiaire.getEmail());
+		String retourMail = mailRepository.sendEmailToEntreprise(mailToEntreprise, stagiaire.getCV());
+		if (retourMail.equals("ko")) {return new ResponseEntity<>(new ResponseMessage("pas de CV !"), HttpStatus.FORBIDDEN);}
+		
+		StagiaireMail mailToStagiaire = new StagiaireMail(stagiaire.getEmail(), offre.getTitre(), requestPostuler.getMotivation(),
+				stagiaire.getPrenom(), offre.getEntreprise().getRaisonSociale());
+		mailRepository.sendEmailToStagiaire(mailToStagiaire);
+		
 		repository.save(offre);
 
-		HTMLMail mailToEntreprise = new HTMLMail(offre.getEntreprise().getEmail(), offre.getTitre(),
-				requestPostuler.getMotivation(), stagiaire.getPrenom(), stagiaire.getName(), stagiaire.getEmail());
-		mailRepository.sendEmailToEntreprise(mailToEntreprise);
-
-		String messageStagiaire = "Vous avez envoyez votre candidature à " + offre.getEntreprise().getEmail() + ": "
-				+ requestPostuler.getMotivation();
-		HTMLMail mailToStagiaire = new HTMLMail(stagiaire.getEmail(), offre.getTitre(), messageStagiaire,
-				stagiaire.getPrenom(), stagiaire.getName(), stagiaire.getEmail());
-		mailRepository.sendEmailToEntreprise(mailToStagiaire);
+		repository.save(offre);
 
 		return new ResponseEntity<>(new ResponseMessage("Vous avez bien postulé à cette offre!"), HttpStatus.OK);
 	}
 
-	public Set<Stagiaire> getPostulants(long id) {
+	public ResponseEntity<?> getPostulants(long id) {
 		
 		Offre offre = repository.findById(id)
 				.orElseThrow(() -> new UsernameNotFoundException("Offre Not Found with -> id : " + id));
+		String userEntreprise = offre.getEntreprise().getUsername();
 
-		Set<Stagiaire> postulants = offre.getStagiaires();
-		
-		return postulants;
+		if (authenticationUser.isValidate(userEntreprise)) {
+			Set<Stagiaire> postulants = offre.getStagiaires();		
+			return new ResponseEntity<>(postulants, HttpStatus.OK);
+		}
+		else {
+			return new ResponseEntity<>(new ResponseMessage("Pas le bon utilisateur !"), HttpStatus.FORBIDDEN);
+		}
 
 	}
 }
